updated functions with zoe's help. need to remodel board.js and boardCards.js after nctions
